split up secondary/primary db classes
