+++ hub/db/__init__.py
@@ -1,1 +1,1 @@
-from .db import HubDB
+from .db import SecondaryDB
